+++ src/base/NextApi.Common/Entity/ISoftDeletableEntity.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace NextApi.Common.Entity
+﻿namespace NextApi.Common.Entity
 {
     /// <summary>
     /// Interface indicates that entity can be soft deleted
@@ -11,28 +9,5 @@ namespace NextApi.Common.Entity
         /// Indicates that entity soft-deleted or not
         /// </summary>
         bool IsRemoved { get; set; }
-    }
-
-    /// <summary>
-    /// Interface represents entity to be logged with string user id property type
-    /// </summary>
-    public interface ILoggedSoftDeletableEntity<TUserIdType> : ISoftDeletableEntity
-    {
-        /// <summary>
-        /// Contains id of user removed this entity last time
-        /// </summary>
-        TUserIdType RemovedById { get; set; }
-
-        /// <summary>
-        /// Contains time when entity last time removed
-        /// </summary>
-        DateTimeOffset? Removed { get; set; }
-    }
-
-    /// <summary>
-    /// Interface represents entity to be logged with string user id property type
-    /// </summary>
-    public interface ILoggedSoftDeletableEntity : ILoggedSoftDeletableEntity<string>
-    {
     }
 }
